Fix stats reporting more than total disk space.

+++ engine/manifest.go
@@ -46,58 +46,60 @@ func (n *Node) SyncManifest() {
 		log.Println(err)
 	}
 
-	// Boot adder subsystem
-	toAdder, err := n.FileAdder()
-	if err != nil {
-		log.Println(err)
-	}
-
-	for file := range files {
-		// If file has been added to the manifest, add it
-		// to the database and check number of times replicated
-		// to determine if file should be replicated locally.
-		if file.Status == "add" {
-			file.Status = "remote"
-
-			// Add file to database.
-			err = n.DB.Add(file)
-			if err != nil {
-				log.Println(err)
-				continue
-			}
-
-			// Send file to adder subsystem
-			toAdder <- file
+	if len(files) > 0 {
+		// Boot adder subsystem
+		toAdder, err := n.FileAdder()
+		if err != nil {
+			log.Println(err)
 		}
 
-		// If file has been deleted from the manifest, remove
-		// it from the database and unpin it from the embedded
-		// IPFS node if necessary.
-		if file.Status == "remove" {
-			result, err := n.DB.Remove(file.ID)
-			if err != nil {
-				log.Println(err)
-				continue
-			}
+		for file := range files {
+			// If file has been added to the manifest, add it
+			// to the database and check number of times replicated
+			// to determine if file should be replicated locally.
+			if file.Status == "add" {
+				file.Status = "remote"
 
-			if result.Status == "local" {
-				err = n.Node.Unpin(result.ID)
+				// Add file to database.
+				err = n.DB.Add(file)
 				if err != nil {
 					log.Println(err)
 					continue
 				}
+
+				// Send file to adder subsystem
+				toAdder <- file
+			}
+
+			// If file has been deleted from the manifest, remove
+			// it from the database and unpin it from the embedded
+			// IPFS node if necessary.
+			if file.Status == "remove" {
+				result, err := n.DB.Remove(file.ID)
+				if err != nil {
+					log.Println(err)
+					continue
+				}
+
+				if result.Status == "local" {
+					err = n.Node.Unpin(result.ID)
+					if err != nil {
+						log.Println(err)
+						continue
+					}
+				}
 			}
 		}
-	}
 
-	// Close adder subsystem
-	close(toAdder)
+		// Close adder subsystem
+		close(toAdder)
 
-	// Clean up IPFS store after syncing manifest.
-	err = n.Node.GC()
-	if err != nil {
-		log.Println(err)
-		return
+		// Clean up IPFS store after syncing manifest.
+		err = n.Node.GC()
+		if err != nil {
+			log.Println(err)
+			return
+		}
 	}
 
 	fmt.Printf("Successfully imported & updated manifest\n")
